move App to root folder
